Api Compras verifica estoque antes de seguir com o processo

+++ apicompras/src/main/kotlin/com/nadir/apicompras/entities/Compra.kt
@@ -7,7 +7,7 @@ import java.math.BigDecimal
 data class Compra (
     val idCliente: Long,
     val idProduto: Long,
-    val qtdItensComprados: Integer,
+    val qtdItensComprados: BigDecimal,
     val valorUnitarioDoItem: BigDecimal,
     val valorTotal: BigDecimal
         )
